+++ Phifty/Web.php
@@ -18,20 +18,6 @@ class Web
         return $html;
     }
 
-
-    /**
-     * @param string[] $assets asset names
-     * @param string   $target   name
-     *
-     * {{ Web.include_assets(['jquery','jquery-ui'], 'page_name')|raw}}
-     */
-    public function include_assets($assetNames, $target = null)
-    {
-        $kernel = kernel();
-        $assetObjs = $kernel->asset->loader->loadAssets($assetNames);
-        return $kernel->asset->render->renderAssets($assetObjs,$target);
-    }
-
     public function langs()
     {
         return kernel()->locale->getLangList();
@@ -40,18 +26,6 @@ class Web
     public function get_result( $resultName )
     {
         $runner = ActionRunner::getInstance();
-
         return $runner->getResult( $resultName );
     }
-
-    public function render_result( $resultName )
-    {
-        $runner = ActionRunner::getInstance();
-        if ( $result = $runner->getResult( $resultName ) ) {
-            $view = new View;
-            $view->result = $result;
-            return $view->render('@CoreBundle/phifty/action_result_box.html');
-        }
-    }
-
 }